Move ProcessMonitor into namespace Services
