+++ Services/MainFormService.cs
@@ -4,7 +4,6 @@ using System.Diagnostics;
 using System.Windows.Forms;
 using Microsoft.Win32;
 using OBSLauncher.Abstractions;
-using OBSLauncher.Domain;
 using OBSLauncher.Validators;
 
 namespace OBSLauncher.Services
